requirements.txt
pdfplumber==0.10.3
pandas==2.1.4
openpyxl==3.1.2
tabula-py==2.9.0
PyPDF2==3.0.1
psycopg2-binary==2.9.9
plotly==5.18.0

